more tests and new test file
